simple_client/requirements.txt
authlib>=1.0.1
flask>=2.2.2
python-dotenv>=0.20.0
requests>=2.28.1